Patch 1.0
Completed basic functionality. The Robot can now find an efficient path to the goal, which is represented in the terminal as a 2D matrix with "X''s at points the robot has been. It does not mark every grid square it passes through, this lets you see the difference in magnitude of movement between points (speed), this can be though of as the robot "checking in" its position at regular time intervals